Automate to prepare release note with next version (#340)
To create release notes, the following steps are necessary, but
currently, they cannot be enforced:

1. Decide on a title.
2. Create a version following semver.
3. Revise the automatically generated release notes.

This change automates the creation of a draft release, forcing the
release creator to complete steps 1 and 2. The version is automatically
determined with a `v` prefix by simply choosing from major/minor/patch.
The title will be a combination of steps 1 and 2.

About 3, drafted note shows actor name, so I believe almost all of us
can realize we need to revise note.(but I know it's not perfect)

### When we run the workflow
When the workflow is executed, a draft like the following will be
created.
![スクリーンショット 2024-11-25 15 55
00](https://github.com/user-attachments/assets/adf62705-55da-4096-a8ac-2bb4bb158fc8)



By revising this and making a release, it will result in the same as
usual.
![スクリーンショット 2024-11-25 15 55
20](https://github.com/user-attachments/assets/1ba38c80-5bdc-4627-8957-a8e9a32ae3ec)