Bye changelog (#415)
In all other bot-sdk repositories, the release note serves as the
changelog. In practice, the release note already contains the same
content as the changelog.
- https://github.com/line/line-bot-sdk-ruby/releases

GitHub's release notes are often automated and serve as a source of
information that tools like Renovate and Dependabot display in PRs.
There is no point in maintaining both. Let's use only the release note.