Add test to check content type header (#464)
This change adds tests for the headers for scenario about downloading.

The behavior where all header names become lowercase when converting the
response header object to a hash likely depends on `Net::HTTP`.

Some libraries including `Net::HTTP` convert uppercase access to
lowercase(In ruby, String#downcase.to_s), but once it's converted to a
hash in bot-sdk-ruby, they don't do this.
However, since HTTP header names are not supposed to be case-sensitive
(regardless of the programming language), I think this behavior is
acceptable.