Add test for stateless channel access token (#465)
`POST /oauth2/v3/token` is somewhat special. Out of the five parameters,
only three need to be filled. For the remaining ones, either specify nil
or don't specify them at all, allowing the SDK user to successfully make
the request.

Since there is no test for this, this change adds it.

Additionally, this change adds testes to ensure that the Authorization
header is not sent by the channel access token client.