Add comments for clients (#468)
This change adds comments for sdk users in code related to the client.

(1) The internal HTTP client is exposed, but this change adds caution
against using it. If a breaking change is made to this, the version will
be raised, ignoring semver.
(2) For the user-facing HTTP client, this change adds comments with
examples using YARD.

Comments on the class methods will be improved in a separate PR.