Add more tests for get followers API (#467)
This change re-wrote some test cases slightly and adds a case where
there is access twice in one scenario.
This is because, in the past, there were cases in other bot SDKs where
making a request twice would cause issues. As we can see from the code,
this doesn't happen in the Ruby SDK, but this test is just to be
safe....