Add test for PUT&POST /v2/bot/audienceGroup/upload/byFile (#463)
About `PUT&POST /v2/bot/audienceGroup/upload/byFile`, 
the value for content-type header must be `multipart/form-data`, and
content-type for the file within the http request's body must be
`text/plain`. This change adds a test for that.
It's somewhat rough, but I don't think this will cause a regression, so
I believe this test is sufficient.