Comment for model (#471)
(WIP) depends on the previous PR
https://github.com/line/line-bot-sdk-ruby/pull/469

This change provides YARD with sufficient information about the
models(classes). Many of the comments created are the same as the types
in RBS, as well as existing descriptions and variable names.

Since YARD uses a different notation than RBS, you can consider this as
an adaptation to that notation, and there should be almost no issues.