Use rbs collection for validation by rbs command  (#479)
In the next PR, it will define the type for the HTTP client and make
slight modifications to the code. In the process, types from other
libraries will be needed. This change is a preparation for that. It
imports the types of the necessary external libraries using the `rbs
validate` command.