Make integer to string for path parameter (#483)
It seems that we cannot pass an `Integer` to `String#gsub`. Since both
`String` and `Integer` have a `#to_s` method, this change ensures to
always call this method for all given parameter values. As a result,
`String#gsub` will always receive a value of type String. This was
discovered thanks to Steep.