Handling Version Loading in Gems Properly (#504)
If the version file is not loaded in the library code but is loaded in
the gemspec file, issues like
(https://github.com/line/line-bot-sdk-ruby/pull/379) load errors or
missing constants can occur.

To be cautious, the gemspec avoids requiring the version file directly.
Instead, it reads the file content. This approach also includes a simple
validation to ensure the version string follows the SemVer format, so
mistakes in the version file can be detected early.

Additionally, this change fixes an actual case where the version file
was not being required properly.