Use require 'line-bot-api' instead of require 'line/bot' (#507)
From v2, we recommend using `require 'line-bot-api'` instead of `require
'line/bot'`. Of course, `require 'line/bot'` will still work, but using
the name that matches the gem is clearer. Additionally, `'line/bot'` is
an internal representation and not something users need to be aware of.

This is not a breaking change. It will still work with `require
'line/bot'`. However, I don't think we need to guarantee this behavior
from v2.

Close https://github.com/line/line-bot-sdk-ruby/issues/495