Missing variable definition in class for RBS (#517)
When using `@variable` within a class, if we don't write its definition
in the RBS file, `@variable.hoge` will be inferred as `untyped`.

As a result, all subsequent values will be considered `untyped`, and
mistakes cannot be detected. (like
https://github.com/line/line-bot-sdk-ruby/issues/499)

To prevent this, this change adds type definitions for variables within
the class. Note steep with strict mode can show hint about this.