Detect unloaded files under lib directory (#506)
This change adds specs to detect files that are not required yet by
`required line-bot-api` or `require line/bot`. However, since RSpec runs
tests in the same process, it heavily depends on the order of test
execution. Additionally, after `require line-bot-api`, we cannot clean
up and then cannot call `require line/bot`.

Therefore, this change makes sure that important tests are run in a
separate process. To avoid having to manually enter commands locally
with test separation in mind, this change describes all the tests in the
Rakefile's test task to hide the details of test execution.