Write actual dependency to line-bot-api.gemspec (#531)
In the `*.gemspec` file, we need to list the files the library depends
on, specifying at least the minimum versions. Additionally, bundled gems
excluded from the standard Ruby install should be explicitly listed,
such as `base64`.
-
https://github.com/ruby/ruby/blob/93afcfcde36581e6f94b69c3f40fd0021f382d70/gems/bundled_gems
-
https://github.com/ruby/ruby/blob/6962f3dc29702ffa68f33d889bab7067b6fe0ca7/lib/bundled_gems.rb#L3-L28

Development dependencies can be written in either the `*.gemspec` or the
`Gemfile`. it may be ok to write them in `Gemfile` as of now.