Clarify response type (#546)
This change clarifies the return type of the function in both YARD and
RBS. Due to issues with steep
(https://github.com/soutaro/steep/issues/1603), this change disables a
useful rule that should normally be enabled.
I verified using RBS runtime tests in advance and chose to ignore steep
errors. As a result, there are currently no additional errors to fix.
There might be unknown patterns that could slip through, but eliminating
this risk would require submitting a patch to steep, which I have
decided against...