Start to run steep in CI (#515)
Using RBS, we were able to detect some minor bugs before release. It
also enabled detection of discrepancies between implementation and RBS.
We have now reduced errors in strict mode to zero!

To automatically catch issues in future changes, let's always verify
this. To achieve this, this change adds the job to run steep in CI.