Sync comment between code and rbs (#556)
Resolve https://github.com/line/line-bot-sdk-ruby/issues/534

When jumping to RBS in the IDE, there are few comments. Having comments
would be helpful. It doesn't have to be YARD; RDoc might be preferable,
but having information is the priority. This change synchronizes with
the code comments. There are no other changes.