Fix unnecessary camelize issue (#567)
fix #562

This solves the problem of camelizing API fields that should not be
camelized.
In short, when using `TextV2` messages, the user can specify their own
key in the `substituion`, but the bug occurs when snake_case is used
there.

This should be set in the request class side, but since request body
specification by hash without request class is also accepted, we have to
deal with this workaround.

This is fine for now because there is only `substitution`, but if a
pattern caramelizes and does not caramelize the same key name in the
future, a major modification or breaking change may be necessary.