Try to use github pages instead of rubydoc.info (#588)
It sometimes takes time to open `rubydoc.info`, and while it is
convenient, this can be a little bit stressful. Since the complete HTML
can be built locally thanks to yard, it can be moved to GitHub Pages.

First, let's test it. When everything works as expected, the following
will be executed in a separate PR:
- Add github actions workflow to deploy pages on release
- Swap the URL (gemspec, readme, GitHub)