Fix namespaces in migration guide (#596)
Some examples are wrong in migration guide. This change fixes them.

https://github.com/line/line-bot-sdk-ruby/releases/tag/v1.30.0 has been
already fixed.

This is detected with this script
(https://gist.github.com/Yang-33/57e571ffecd6dd755c2d1445de1abd2d). This
is a makeshift script and might detect incorrectly. Since maintenance is
not straightforward, I'll avoid adding it as a spec for now.